pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-tailor"
version = "0.1.0"
description = "AI-powered resume tailoring system using Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "tavily-python>=0.5.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "streamlit>=1.32.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
resume-tailor = "resume_tailor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_tailor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
